requirements.txt
# Editable install of hackaton
-e .
accelerate>=0.20.0
click
flask>=2.0.0
flask-cors>=3.0.10
ipympl
ipython
jupyter
jupyterlab
matplotlib
nbdime
nbstripout
num2words
numpy
pandas
pillow
pre-commit
pytest
pytest-cov
recommonmark
safetensors>=0.3.1
scikit-learn
scipy
seaborn
sphinx
tensorflow
tf-keras
torch
torch-vision
tox
tqdm
transformers


# Uncomment if needed
# statsmodels
# sktime
# tslearn
# xgboost
# catboost
# lightgbm
# pytorch
# keras
# spacy
# optuna
# dask
# snakeviz
# icecream
# jax
# numpyro
# funsor
# neptune-client
# neptune-contrib
